Refactoring and creating tests

+++ src/main/kotlin/com/matera/restserver/service/EmployeeService.kt
@@ -4,63 +4,54 @@ import com.matera.restserver.dto.CreateEmployeeDTO
 import com.matera.restserver.exception.EntityNotFoundException
 import com.matera.restserver.model.Employee
 import com.matera.restserver.repository.EmployeeRepository
-import com.matera.restserver.util.Messages
 import org.springframework.stereotype.Service
+import java.util.UUID
 import java.util.stream.Collectors
 
 @Service
 open class EmployeeService(private val employeeRepository: EmployeeRepository) {
+    fun create(dto: CreateEmployeeDTO): Employee = employeeRepository.save(
+        Employee(
+            firstName = dto.firstName,
+            middleName = dto.middleName,
+            lastName = dto.lastName,
+            status = dto.status,
+            dateOfEmployment = dto.dateOfEmployment,
+            dateOfBirth = dto.dateOfBirth
+        )
+    )
 
-    fun create(dto: CreateEmployeeDTO): Long {
-        return employeeRepository.save(
-            Employee(
-                firstName = dto.firstName,
-                middleName = dto.middleName,
-                lastName = dto.lastName,
-                status = dto.status,
-                dateOfEmployment = dto.dateOfEmployment,
-                dateOfBirth = dto.dateOfBirth
-            )
-        ).id
-    }
-
-    fun find(id: Long): Employee {
+    fun find(id: UUID): Employee? {
         return employeeRepository.findByIdAndStatus(id, ACTIVE)
-            ?: throw EntityNotFoundException("exceptions.default.entitynotfoundexception")
     }
 
-    fun findAll(): List<Employee> {
-        val employees = employeeRepository.findByStatus(ACTIVE)
-        if (employees.isEmpty()) {
-            throw EntityNotFoundException(Messages.getMessage("business.employee.employeesnotfound"))
-        }
-        return employees
-    }
+    fun findAll(): List<Employee> = employeeRepository.findByStatus(ACTIVE)
 
-    fun delete(id: Long): Employee {
+    fun delete(id: UUID): Employee? {
         val employee = employeeRepository.findByIdAndStatus(id, ACTIVE)
-            ?: throw EntityNotFoundException("exceptions.default.entitynotfoundexception")
-        employeeRepository.deleteById(id)
+        if (employee != null) employeeRepository.deleteById(id)
         return employee
     }
 
-    fun update(dto: Employee) {
-        val employee = employeeRepository.findByIdAndStatus(dto.id, ACTIVE)
+    fun update(dto: Employee): Employee {
+        employeeRepository.findByIdAndStatus(dto.id, ACTIVE)
             ?: throw EntityNotFoundException("exceptions.default.entitynotfoundexception")
-        employeeRepository.save(
-            Employee(
-                id = dto.id,
-                firstName = dto.firstName,
-                middleName = dto.middleName,
-                lastName = dto.lastName,
-                status = dto.status,
-                dateOfEmployment = dto.dateOfEmployment,
-                dateOfBirth = dto.dateOfBirth
-            )
+        val employee = Employee(
+            id = dto.id,
+            firstName = dto.firstName,
+            middleName = dto.middleName,
+            lastName = dto.lastName,
+            status = dto.status,
+            dateOfEmployment = dto.dateOfEmployment,
+            dateOfBirth = dto.dateOfBirth
         )
+        employeeRepository.save(
+            employee
+        )
+        return employee
     }
 
-    fun create(dto: List<CreateEmployeeDTO>?) {
+    fun create(dto: List<CreateEmployeeDTO>?): List<Employee> {
         val entities = dto!!.stream().map { e: CreateEmployeeDTO ->
             val d = Employee(
                 firstName = e.firstName,
@@ -72,7 +63,7 @@ open class EmployeeService(private val employeeRepository: EmployeeRepository) {
             )
             d
         }.collect(Collectors.toList())
-        employeeRepository.saveAll(entities)
+        return employeeRepository.saveAll(entities)
     }
 
     companion object {
